+++ src/Pim/Bundle/CatalogBundle/Entity/Repository/AssociationRepository.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\CatalogBundle\Entity\Repository;
 
 use Pim\Bundle\CatalogBundle\Doctrine\EntityRepository;
 use Pim\Bundle\CatalogBundle\Entity\AssociationType;
+use Pim\Bundle\CatalogBundle\Model\AssociationRepositoryInterface;
 
 /**
  * Association repository
@@ -12,14 +13,11 @@ use Pim\Bundle\CatalogBundle\Entity\AssociationType;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AssociationRepository extends EntityRepository implements ReferableEntityRepositoryInterface
+class AssociationRepository extends EntityRepository implements AssociationRepositoryInterface,
+ ReferableEntityRepositoryInterface
 {
     /**
-     * Return the number of Associations for a specific association type
-     *
-     * @param AssociationType $associationType
-     *
-     * @return mixed
+     * {@inheritdoc}
      */
     public function countForAssociationType(AssociationType $associationType)
     {