+++ src/Pim/Bundle/CatalogBundle/Validator/Constraints/UniqueVariantAxisValidator.php
@@ -156,7 +156,6 @@ class UniqueVariantAxisValidator extends ConstraintValidator
      * Add violation to the executioncontext
      *
      * @param Constraint $constraint
-     * @param string     $productLabel
      * @param string     $variantLabel
      * @param string     $values
      */